// src/api/generated/models/MenuAnalyticsOverview.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Menu analytics overview with key metrics and trends
 */
export type MenuAnalyticsOverview = {
    business_id: string;
    period: string;
    total_menu_items: number;
    popular_items: number;
    average_rating: number;
    total_categories: number;
    items_growth: number;
    rating_growth: number;
    categories_growth: number;
    popularity_growth: number;
    performance_score: number;
    last_updated: string;
    trends_included?: boolean;
};